Use ordered hashes to store windows and tabs
The issue was in Ruby 1.8, AbstractCore#last_open_window was not always
returning the last open window, because it calls keys.last on the windows hash,
which could be a random key. This meant that if your termfile opened a window
with multiple tabs, a random number of those tabs would be in the current window
and the rest would be in the new window.

The solution is to use ActiveSupport's OrderedHash implementation to store
windows and tabs so hopefully we don't run into this problem (and other related
problems) again.